feat: add review

+++ src/main/kotlin/com/example/springbackend/matjib/infrastructure/MatjibEntity.kt
@@ -19,5 +19,5 @@ class MatjibEntity(
 
 class ScoreEntity(
     val userId: String,
-    val score: Int
+    val score: Double
 )
