+++ src/main/kotlin/com/example/springbackend/matjib/infrastructure/router/dtos/MatjibRequest.kt
@@ -3,7 +3,7 @@ package com.example.springbackend.matjib.infrastructure.router.dtos
 class MatjibRequest(
     val longitude: Double,
     val latitude: Double,
-    val score: Int,
+    val score: Double,
     val name: String,
     val description: String,
     val address: String
